chore: detached widgets drop arena
