+++ app/client/src/layoutSystems/anvil/canvasArenas/styles.css
@@ -1,0 +1,21 @@
+.detached-widgets-drop-overlay {
+  background-color: rgba(255, 229, 254, 0.9);
+}
+
+.detached-widgets-drop-overlay-content {
+  background-color: #ffe5fe !important;
+  height: clamp(40vh, 60vh, 80vh);
+  padding: 0px !important;
+}
+.detached-widgets-drop-overlay-content:hover {
+  border: 1px dotted #ce01d3;
+  border-radius: 6px;
+}
+.detached-widgets-drop-overlay-content > div {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex: fit-content;
+  justify-content: center;
+  align-items: center;
+}
